Amélioration
Extract DirectusApiError to expose this object.
