+++ lib/directus_api_manager.dart
@@ -8,3 +8,4 @@ export 'src/model/directus_file.dart';
 export 'src/model/directus_login_result.dart';
 export 'src/model/directus_item.dart';
 export 'src/model/directus_item_creation_result.dart';
+export 'src/model/directus_api_error.dart';
